Finalise before latest PR review
Split out the PlaySound function into PlaySound and PlayMusic and now instead of passing a bus, it automatically selects the .ogg/.mp4 from the respective audio folder (also created beginnings of directory setup for audio) and plays on the correct tres bus. Also made it so that sounds and music can now be played over one another.

Will just need review and lengthy process of resolving conflicts (a lot has evolved since initial PR) and making this code fit in with the overall picture.