Began creating combat module and spawner module.
Was able to create the combat module that can send a custom signal to the spawner module which then executes the spawn method.
No functionality added, only tested signal. Plan is to update the signal to send a godotobject as a payload that includes the desired parameters. Further instructions on git issue.